resetlogout button
