+++ Code/Client/admin/src/components/molecules/Header.js
@@ -1,8 +1,12 @@
 import React from "react";
+import Swal from 'sweetalert2';
+import { useDispatch } from "react-redux";
 import { LiaBarsSolid } from "react-icons/lia";
 import styles from "./Header.module.css";
 import { useLocation } from "react-router-dom";
-import Button from '../atoms/buttons/Button'
+import Button from '../atoms/buttons/Button';
+import { setSignedOut } from "../../redux/containers/auth/actions";
+
 const capitalizeFirstLetter = (str) => {
   return str.charAt(0).toUpperCase() + str.slice(1);
 };
@@ -11,8 +15,10 @@ const replaceHyphensWithSpaces = (str) => {
   return str.replace(/-/g, " ");
 };
 
-const TopHeader = ({ openSidebar, setOpenSidebar,haveButton,ButtonText,onClick }) => {
+const TopHeader = ({ openSidebar, setOpenSidebar, haveButton, ButtonText, onClick }) => {
   const location = useLocation();
+
+  const dispatch = useDispatch();
 
   const getPageName = () => {
     const path = location.pathname;
@@ -24,6 +30,27 @@ const TopHeader = ({ openSidebar, setOpenSidebar,haveButton,ButtonText,onClick }
     }
   };
 
+  const logout = () => {
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You will be logged out of this session!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, logout!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(setSignedOut());
+        Swal.fire(
+          'Logged Out!',
+          'You have been logged out.',
+          'success'
+        );
+      }
+    });
+  }
+
   return (
     <>
       <section className={styles.topHeader}>
@@ -32,17 +59,17 @@ const TopHeader = ({ openSidebar, setOpenSidebar,haveButton,ButtonText,onClick }
             color="#000000"
             onClick={() => setOpenSidebar(!openSidebar)}
             size={24}
-          /> 
+          />
         </div>
-<div className={styles.content}>
-        <div className={styles.navbar}>
-          <h2>{getPageName()}</h2>
-        </div>
-        <div>
-        {haveButton &&
-        <Button btnText={ButtonText || "Add Text"} primary btnClick={onClick}/>
-        }<Button/>
-        </div>
+        <div className={styles.content}>
+          <div className={styles.navbar}>
+            <h2>{getPageName()}</h2>
+          </div>
+          <div>
+
+            <button className={styles.btn} onClick={() => logout()}>logout</button>
+
+          </div>
         </div>
       </section>
     </>
