+++ Code/Client/admin/src/components/molecules/Header.module.css
@@ -33,6 +33,21 @@
     margin-right: 24px;
     flex-direction: row;
 }
+
+.btn {
+    background-color:  #7E22CE;
+    padding: 0.70rem 1rem;
+    text-transform: uppercase;
+    bottom: 20px;
+    align-self: center;
+    color: #fff;
+    font-weight: 700;
+    font-size: 14px;
+    width: 100%;
+    border-radius: 10px;
+    text-align: center;
+    cursor: pointer;
+}
 /* Show hamburger menu on small screens */
 @media screen and (max-width: 1068px) {
     .hamburger {
